+++ tests/common/mod.rs
@@ -1,6 +1,7 @@
 mod matrix;
-mod run_case;
+mod run_case_closure;
+mod run_case_function;
 
 pub use matrix::float_matrix_comparison;
-pub use run_case::run_test_case_fd;
-pub use run_case::run_test_case_jac;
+pub use run_case_closure::{run_closure_case_fd, run_closure_case_jac};
+pub use run_case_function::{run_function_case_fd, run_function_case_jac};