+++ tests/solver/stationary_newton/broyden.rs
@@ -4,13 +4,13 @@ use util::test_cases::broyden1965::*;
 
 use nrf::solver::{QuasiNewtonMethod, ResolutionMethod};
 
-use crate::common::{run_test_case_fd, run_test_case_jac};
+use crate::common::{run_function_case_fd, run_function_case_jac};
 
 #[test]
 fn broyden_case5_fd() {
     let problem_size = 5;
     let damping = false;
-    run_test_case_fd(
+    run_function_case_fd(
         problem_size,
         broyden1965_case5,
         init_broyden1965_case5(),
@@ -24,7 +24,7 @@ fn broyden_case5_fd() {
 fn broyden_case5_jac() {
     let problem_size = 5;
     let damping = false;
-    run_test_case_jac(
+    run_function_case_jac(
         problem_size,
         broyden1965_case5,
         broyden1965_case5_jac,
@@ -39,7 +39,7 @@ fn broyden_case5_jac() {
 fn broyden_case6_fd() {
     let problem_size = 5;
     let damping = false;
-    run_test_case_fd(
+    run_function_case_fd(
         problem_size,
         broyden1965_case6,
         init_broyden1965_case6(),
@@ -53,7 +53,7 @@ fn broyden_case6_fd() {
 fn broyden_case6_jac() {
     let problem_size = 5;
     let damping = false;
-    run_test_case_jac(
+    run_function_case_jac(
         problem_size,
         broyden1965_case6,
         broyden1965_case6_jac,
@@ -68,7 +68,7 @@ fn broyden_case6_jac() {
 fn broyden_case7_fd() {
     let problem_size = 10;
     let damping = false;
-    run_test_case_fd(
+    run_function_case_fd(
         problem_size,
         broyden1965_case7,
         init_broyden1965_case7(),
@@ -82,7 +82,7 @@ fn broyden_case7_fd() {
 fn broyden_case7_jac() {
     let problem_size = 10;
     let damping = false;
-    run_test_case_jac(
+    run_function_case_jac(
         problem_size,
         broyden1965_case7,
         broyden1965_case7_jac,
@@ -97,7 +97,7 @@ fn broyden_case7_jac() {
 fn broyden_case8_fd() {
     let problem_size = 20;
     let damping = false;
-    run_test_case_fd(
+    run_function_case_fd(
         problem_size,
         broyden1965_case8,
         init_broyden1965_case8(),
@@ -111,7 +111,7 @@ fn broyden_case8_fd() {
 fn broyden_case8_jac() {
     let problem_size = 20;
     let damping = false;
-    run_test_case_jac(
+    run_function_case_jac(
         problem_size,
         broyden1965_case8,
         broyden1965_case8_jac,
@@ -126,7 +126,7 @@ fn broyden_case8_jac() {
 fn broyden_case9_fd() {
     let problem_size = 2;
     let damping = false;
-    run_test_case_fd(
+    run_function_case_fd(
         problem_size,
         broyden1965_case9,
         init_broyden1965_case9(),
@@ -140,7 +140,7 @@ fn broyden_case9_fd() {
 fn broyden_case9_jac() {
     let problem_size = 2;
     let damping = false;
-    run_test_case_jac(
+    run_function_case_jac(
         problem_size,
         broyden1965_case9,
         broyden1965_case9_jac,
@@ -155,7 +155,7 @@ fn broyden_case9_jac() {
 //#[test]
 //fn broyden_case10_fd() {
 //    let problem_size = 2;
-//    run_test_case_fd(
+//    run_function_case_fd(
 //        problem_size,
 //        broyden1965_case10,
 //        init_broyden1965_case10(),
@@ -168,7 +168,7 @@ fn broyden_case9_jac() {
 //#[test]
 //fn broyden_case10_jac() {
 //    let problem_size = 2;
-//    run_test_case_jac(
+//    run_function_case_jac(
 //        problem_size,
 //        broyden1965_case10,
 //        broyden1965_case10_jac,